requirements.txt
-i https://pypi.org/simple
aiohappyeyeballs==2.5.0; python_version >= '3.9'
aiohttp==3.11.13; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
aiosqlite==0.21.0; python_version >= '3.9'
alembic==1.15.1; python_version >= '3.9'
anyio==4.8.0; python_version >= '3.9'
apache-airflow==2.10.5; python_version < '3.13' and python_full_version >= '3.8.1'
apache-airflow-providers-common-compat==1.5.0; python_version ~= '3.9'
apache-airflow-providers-common-io==1.5.0; python_version ~= '3.9'
apache-airflow-providers-common-sql==1.23.0; python_version ~= '3.9'
apache-airflow-providers-fab==1.5.3; python_version ~= '3.9'
apache-airflow-providers-ftp==3.12.2; python_version ~= '3.9'
apache-airflow-providers-http==5.2.0; python_version ~= '3.9'
apache-airflow-providers-imap==3.8.2; python_version ~= '3.9'
apache-airflow-providers-smtp==2.0.0; python_version ~= '3.9'
apache-airflow-providers-sqlite==4.0.0; python_version ~= '3.9'
apispec[yaml]==6.8.1; python_version >= '3.9'
argcomplete==3.6.0; python_version >= '3.8'
asgiref==3.8.1; python_version >= '3.8'
attrs==25.1.0; python_version >= '3.8'
babel==2.17.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
cachelib==0.13.0; python_version >= '3.8'
certifi==2025.1.31; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
clickclick==20.10.2
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
colorlog==6.9.0; python_version >= '3.6'
configupdater==3.2; python_version >= '3.6'
connexion[flask]==2.14.2; python_version >= '3.6'
cron-descriptor==1.4.5
croniter==6.0.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
cryptography==44.0.2; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
deprecated==1.2.18; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dill==0.3.9; python_version >= '3.8'
dnspython==2.7.0; python_version >= '3.9'
email-validator==2.2.0; python_version >= '3.8'
flask==2.2.5; python_version >= '3.7'
flask-appbuilder==4.5.3; python_version ~= '3.7'
flask-babel==2.0.0
flask-caching==2.3.1; python_version >= '3.8'
flask-jwt-extended==4.7.1; python_version >= '3.9' and python_version < '4'
flask-limiter==3.10.1; python_version >= '3.9'
flask-login==0.6.3; python_version >= '3.7'
flask-session==0.5.0; python_version >= '3.7'
flask-sqlalchemy==2.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask-wtf==1.2.2; python_version >= '3.9'
frozenlist==1.5.0; python_version >= '3.8'
fsspec==2025.3.0; python_version >= '3.8'
google-re2==1.1.20240702; python_version ~= '3.8'
googleapis-common-protos==1.69.1; python_version >= '3.7'
grpcio==1.70.0; python_version >= '3.8'
gunicorn==23.0.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
importlib-metadata==8.5.0; python_version >= '3.8'
inflection==0.5.1; python_version >= '3.5'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
lazy-object-proxy==1.10.0; python_version >= '3.8'
limits==4.1; python_version >= '3.9'
linkify-it-py==2.0.3; python_version >= '3.7'
lockfile==0.12.2
mako==1.3.9; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.26.1; python_version >= '3.9'
marshmallow-oneofschema==3.1.1; python_version >= '3.8'
marshmallow-sqlalchemy==0.28.2; python_version >= '3.7'
mdit-py-plugins==0.4.2; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
methodtools==0.4.7
more-itertools==10.6.0; python_version >= '3.9'
multidict==6.1.0; python_version >= '3.8'
opentelemetry-api==1.30.0; python_version >= '3.8'
opentelemetry-exporter-otlp==1.30.0; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-common==1.30.0; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-grpc==1.30.0; python_version >= '3.8'
opentelemetry-exporter-otlp-proto-http==1.30.0; python_version >= '3.8'
opentelemetry-proto==1.30.0; python_version >= '3.8'
opentelemetry-sdk==1.30.0; python_version >= '3.8'
opentelemetry-semantic-conventions==0.51b0; python_version >= '3.8'
ordered-set==4.1.0; python_version >= '3.7'
packaging==24.2; python_version >= '3.8'
pathspec==0.12.1; python_version >= '3.8'
pendulum==3.0.0; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
prison==0.2.1
propcache==0.3.0; python_version >= '3.9'
protobuf==5.29.3; python_version >= '3.8'
psutil==7.0.0; python_version >= '3.6'
pycparser==2.22; python_version >= '3.8'
pygments==2.19.1; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
python-daemon==3.1.2; python_version >= '3.7'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
python-nvd3==0.16.0
python-slugify==8.0.4; python_version >= '3.7'
pytz==2025.1
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rfc3339-validator==0.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rich==13.9.4; python_full_version >= '3.8.0'
rich-argparse==1.7.0; python_version >= '3.8'
rpds-py==0.23.1; python_version >= '3.9'
setproctitle==1.3.5; python_version >= '3.8'
setuptools==78.1.0; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==1.4.54; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlalchemy-jsonfield==1.0.2; python_full_version >= '3.7.0'
sqlalchemy-utils==0.41.2; python_version >= '3.7'
sqlparse==0.5.3; python_version >= '3.8'
tabulate==0.9.0; python_version >= '3.7'
tenacity==9.0.0; python_version >= '3.8'
termcolor==2.5.0; python_version >= '3.9'
text-unidecode==1.3
time-machine==2.16.0; python_version >= '3.9'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
uc-micro-py==1.0.3; python_version >= '3.7'
universal-pathlib==0.2.6; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'
werkzeug==2.2.3; python_version >= '3.7'
wirerope==1.0.0
wrapt==1.17.2; python_version >= '3.8'
wtforms==3.2.1; python_version >= '3.9'
yarl==1.18.3; python_version >= '3.9'
zipp==3.21.0; python_version >= '3.9'
